update .env file sourcing
This makes some changes to how `.env` files are read by the various
build scripts that we have in here:

- there is now a `common.sh` file where there is one function called
  `load_env` which sources the `.env` file
- the function will filter out lines starting with `#`, enabling the use
  of comments in `.env` files
- the function also accepts a `--require-dot-env` argument which will
  cause it to error if a `.env` file is not present

We use `.env` files to set environment variables for local development,
but for production builds these are set in the environment already. So
while we always want to grab any variables in a `.env` file, we don't
always want to fail the build (or dev environment start script) if such
a file is not present.

closes #178
pr #180