switch to building the JS files with Typescript
This sets up `ts-loader` for Webpack and converts most of the JS files
over to Typescript (especially all the files for the frontend search). I
tried not to touch any logic at all in doing so. Many files are just
renamed from `.js` to `.ts` without any code changes.

Some files still have a number of type errors, but these are silenced
with a @ts-nocheck at the top of the file. We should circle back and
remove these over time.

closes #324
pr #335