fix usage of aria-live on the search page
There is more information on the linked issue (#378).

Basically we mis-used `aria-live` in a way that made the search page
pretty terrible for using a screenreader. In short, all of the changes
within a div marked `aria-live="polite"` are read out when they change,
and we were wrapping both the search facets and the search results with
that, so every time the results changed the screenreader would read out
all that stuff.

Instead, a recommendation for dynamic search UIs like ours is to only
wrap the result count with `aria-live="polite"`, so that when the search
results update there is a notification to the user but they aren't
deluged with a ton of stuff.

This commit changes our usage so that only the result count is read out.

closes #378
pr #380