fix handling of role="search"
This commit contains two changes to improve accessibility of the search
input:

- sets `role="search"` on the `form` tag which wraps the search input
- sets `type="search"` on the `input` tag itself

These two changes should signal to a screenreader user that the search
box is, well, a search box, and not just a random text field. Setting
`role="search"` on the form should also allow that to be recognized as a
functional region by screenreaders.

closes #379
pr #381