add a data template for creating a course content map
This adds a new json template which renders a map from UUID to relative
link for all pages. We need this in order to render resource
collections, specifically in order to construct URLs for the `data.json`
files for those pieces of content.

More detail on this:

In order to fetch a data.json file in ocw-www we need to construct the
URL for that file. Consider the example URL:

https://ocwnext-rc.odl.mit.edu/courses/16-885j-aircraft-systems-engineering-fall-2004/resources/aircraftpersp-1/

This URL depends on three pieces of information:

1. the baseurl (`https://ocwnext-rc.odl.mit.edu`)
2. the website name (`16-885j-aircraft-systems-engineering-fall-2004`)
3. the relative path for the piece of content (`/resources/aircraftpersp-1/`)

The data we’re storing in Studio for resources gives us piece no. 2, and
we have piece no. 1 because it is set via an environment variable in the
build environment for Hugo. The other piece of information we store in
the resource collection in Studio is the content UUID, because that is
what we use as a URL param for the WebsiteContent APIs in Studio (and
just what we lean on generally for uniquely identifying content).

This commit makes a change so that courses now output a mapping (a JSON
file) from content UUID -> relative path, which will let us get piece
no. 3 given nos. 1 and 2 and a UUID. The JSON file will look something
like this:

```json
{
  "013ea582-8958-24c8-61df-cd7bf1e6526e": "/resources/mit18_s997f11_exercise_17/data.json",
  "01d3a55d-eacc-1672-2d62-09afbfd9ee44": "/resources/mit18_s997f11_exercise_2/data.json",
  "01e581a8-1cba-3b53-0d29-7eb44635a452": "/resources/lecture-3-using-files/data.json",
  "01f41609-eedf-c2d7-11bd-02593bab20d4": "/resources/18-s997f11_unit4_img2/data.json",
  "0523856e-c9fc-9e71-66c9-d2d9aecf63e3": "/pages/vectorization/complex-numbers/data.json",
  "072d4e9f-65ad-773a-5386-2ee2572818f7": "/resources/18-s997f11_unit3_img4/data.json",
  "08cfb41f-1f5e-b1e9-bb10-ba9eae5b1b45": "/pages/library/projects/data.json"
}
```

It will be found at the path `/content_map.json`.

part of #279
pr #400