add support for rendering course collections
A collection is a list-of-lists. This adds basic support for rendering
the description and the course lists within such a collection.

We also recently renamed what we used to call 'course collectins' to be
called 'course lists'. This commit also changes component and variable
names in line with that.

pr #540