Add env.ts + replace dotenv-cli with 'with-env'
env.ts: this is the source-of-truth for accessing environment variables within JS + TS. Think of it like `settings.py` in a Django application. It responsible for (1) loading environment variables from `.env`, (2) setting defaults for optional environment variables, and (3) simple validation (e.g., ports should be numbers within a certain range). This also makes types-safe access much easier: All values on process.env have type `string | undefined`, whereas the values of `import { env } from "./env"` are not nil.

The trouble with `env.ts` is that it is a typescript file, and hence only available in other typescript files. Namely, not available for use in Hugo templates. Enter `with-env.ts`.

with-env.ts: This is a simple CLI utility that runs a shell command with environment variables from `env.ts`. In other words, it is like dotenv-cli except reading from `env.ts` rather than `.env`. This gets us the aforementioned benefits of `env.ts`.

My expectation is that `env.ts` and `with-env` is that these will mostly be used for development. In production, environment variables are already present in the environment and do not need to be loaded from a file.

Because I expect this to be primarily used in development, I have set most of the defaults to be `devDefaults`, i.e., only used in non-production environments.